refactor: remove /analytics/trends endpoint
