--- Pdmt.Api/Dto/Analytics/TrendGranularity.cs
@@ -1,3 +1,0 @@
-namespace Pdmt.Api.Dto.Analytics;
-
-public enum TrendGranularity { Week, Month }
