+++ pdmt-web/src/api/analytics.ts
@@ -1,5 +1,5 @@
 import { apiGet } from "./client";
-import type { CalendarWeekDto, WeeklySummaryDto, TrendPeriodDto } from "./types";
+import type { CalendarWeekDto, WeeklySummaryDto} from "./types";
 
 export const getCalendarWeek = (weekOf: string, signal?: AbortSignal) =>
   apiGet<CalendarWeekDto>(
@@ -10,11 +10,5 @@ export const getCalendarWeek = (weekOf: string, signal?: AbortSignal) =>
 export const getWeeklySummary = (weekOf: string, signal?: AbortSignal) =>
   apiGet<WeeklySummaryDto>(
     `/api/analytics/weekly-summary?weekOf=${encodeURIComponent(weekOf)}`,
-    signal,
-  );
-
-export const getTrends = (from: string, to: string, signal?: AbortSignal) =>
-  apiGet<TrendPeriodDto[]>(
-    `/api/analytics/trends?from=${encodeURIComponent(from)}&to=${encodeURIComponent(to)}&period=week`,
     signal,
   );
